prep: initial code for understanding async/await
As preparation for a new post, this change adds code that shows 3
different async functions and an equivalent manually implemented
`Future`.

Additionally, a top level Cargo.toml manifest has been added so that
this whole repository acts as a workspace. This makes it easier to write
and run the example code.