post: how I finally understood async/await in Rust (part 1)
This is part 1 of what was going to just be a single post on
understanding async/await in Rust. However, this post is already too
long and I probably want another 3 posts to cover everything I wanted
to.

The post is forward dated to tomorrow, as I won't publish it until then
(after sleeping, not in 20 minutes).