fix: remove "text" label for code blocks
Zola (or whatever it's using for syntax highlighting) doesn't like
blocks which are labled as text, such as:

```text
This is just text, don't highlight it.
```

It doesn't cause any issues (as no highlighting is what I want), but
`zola serve` and `zola build` emit warnings.  So I've removed them from
the understanding async/await part 1 post to remove the warnings.