draft: understanding async/await 3
Further work on the article, we have an implementation of the future
now. There were some tricks involved in this implementation (and some
unsafe) and I needed help to get it to compile.

I also changed the async function implementation to not use sleep, but
instead just yield. This has the effect of simplifying the future
substantially.