draft: understanding async/await 4
This post is the 4th (and final planned) part of my series on
understanding async/await in Rust.

The post goes over a case of writing an async channel, specifically a
multi-producer multi-consumer (MPMC) channel. This is a resource that
could not be written with only async/await code, instead manual future
implementations are needed.

So far the draft includes the introduction and a description of what
message passing channels are and some of the types.