post: understanding async/await part 4
This is the completed version of the post (until someone gives
feedback). It includes all the code (with some links this time).

I also added a (very short) README to the understanding-async-await
crate in the repo workspace. This should be extended with links to the
blog posts and instructions on running all the examples.