draft: tracing tokio with traces
Added an explanation of the code. We're up to seeing some traces from
tokio now.

Also added a writer which can be used to convert ANSI output from the
`fmt` layer in tracing-subscriber to HTML so it can be copy-pasted
straight into the blog post.

A little bit of CSS tweaking was added to support these bespoke code
examples.